requirements.txt
fastapi
uvicorn
python-dotenv
motor
httpx
pyjwt
passlib[bcrypt]
starlette
pydantic[email]